+++ src/main/java/Ui.java
@@ -1,0 +1,12 @@
+import java.util.Scanner;
+
+public class Ui {
+
+    public void sayhi() {
+        System.out.println("Hello! I'm LimBay.\nWhat can I do for you?\n");
+    }
+
+    public void saybye() {
+        System.out.println("Bye. Hope to see you again soon!");
+    }
+}
